3D Engine  + Removed file dialog macros  + Updated settings file parser and began implementing it into the engine
git-svn-id: file:///e/SVN/3d_engine@49 463a7de7-ff88-454d-8a36-9b7cb3bdd095

+++ Engine/Source/main.cpp
@@ -1,12 +1,10 @@
-//95.4002 Wilf LaLonde: Student game engine.
-
 #include "includes.all"
 
 char * gameName = "3D Game Engine";
 int mouseSensitivity = 5;
 bool fullscreen = false;
-long initialScreenWidth = 1200;
-long initialScreenHeight = 720;
+int initialScreenWidth = 1200;
+int initialScreenHeight = 720;
 
 void exitGame();
 
@@ -16,8 +14,8 @@ enum MenuItem {DifficultyEasy, DifficultyChallenging, DifficultyImpossible,
 MenuItem difficulty; MenuItem displayMode;
 
 bool wireframe = false;
-long screenWidth;
-long screenHeight;
+int screenWidth;
+int screenHeight;
 char filename [256];
 
 #define escapeCharacter 27
